Moved a couple of modules out of dl_core
