+++ lib/dl_cache_engine/pyproject.toml
@@ -13,6 +13,7 @@ readme = "README.md"
 attrs = ">=22.2.0"
 python = ">=3.10, <3.13"
 datalens-app-tools = {path = "../dl_app_tools"}
+datalens-utils = {path = "../dl_utils"}
 datalens-constants = {path = "../dl_constants"}
 datalens-model-tools = {path = "../dl_model_tools"}
 lz4 = "==4.3.2"
