+++ lib/dl_compeng_pg/pyproject.toml
@@ -13,13 +13,15 @@ readme = "README.md"
 aiopg = ">=1.4.0"
 asyncpg = ">=0.29.0"
 attrs = ">=22.2.0"
-python = ">=3.10, <3.13"
-sqlalchemy = ">=1.4.46, <2.0"
-datalens-sqlalchemy-postgres = {path = "../dl_sqlalchemy_postgres"}
+datalens-api-commons = {path = "../dl_api_commons"}
 datalens-cache-engine = {path = "../dl_cache_engine"}
 datalens-constants = {path = "../dl_constants"}
 datalens-connector-postgresql = {path = "../dl_connector_postgresql"}
 datalens-core = {path = "../dl_core"}
+datalens-sqlalchemy-postgres = {path = "../dl_sqlalchemy_postgres"}
+datalens-utils = {path = "../dl_utils"}
+python = ">=3.10, <3.13"
+sqlalchemy = ">=1.4.46, <2.0"
 
 [tool.poetry.plugins."dl_core.data_processor_plugins"]
 compeng_aiopg = "dl_compeng_pg.compeng_aiopg.plugin:AiopgCompengPlugin"
