+++ lib/dl_core/dl_core/data_processing/types.py
@@ -1,11 +1,7 @@
 from __future__ import annotations
 
-from dl_constants.types import (
-    TBIDataRow,
-    TJSONExt,
-)
-from dl_core.data_processing.streaming import AsyncChunkedBase
+from dl_constants.types import TBIDataRow
+from dl_utils.streaming import AsyncChunkedBase
 
 
 TValuesChunkStream = AsyncChunkedBase[TBIDataRow]  # wrapper of async iterable of chunks of rows
-TJSONExtChunkStream = AsyncChunkedBase[TJSONExt]  # wrapper of async iterable of chunks of jsonables
