+++ lib/dl_core/dl_core_tests/db/compeng/test_pg_op_exec_adapter.py
@@ -20,10 +20,10 @@ from dl_compeng_pg.compeng_pg_base.exec_adapter_base import PostgreSQLExecAdapte
 from dl_constants.enums import UserDataType
 from dl_core.data_processing.cache.utils import CompengOptionsBuilder
 from dl_core.data_processing.processing.context import OpExecutionContext
-from dl_core.data_processing.streaming import AsyncChunked
 from dl_core.services_registry.top_level import ServicesRegistry
 from dl_core.utils import make_id
 from dl_core_tests.db.base import DefaultCoreTestClass
+from dl_utils.streaming import AsyncChunked
 
 
 async def get_active_queries(pg_adapter: PostgreSQLExecAdapterAsync) -> List[Dict[str, Any]]:
